Chat groups and members separated in own components
