+++ client/src/components/ChatGroupsSideBar/index.css
@@ -1,0 +1,36 @@
+.chat-sidebar {
+	background: var(--color-b);
+	color: #fff;
+	padding: 20px 10px 60px;
+	overflow-y: scroll;
+	min-width: 200px;
+	scrollbar-width: none;
+	border: 3px solid white;
+}
+
+h2 {
+	font-size: 20px;
+	background: rgba(10, 10, 10, 0.1);
+	padding: 10px;
+	margin-bottom: 10px;
+	text-align: center;
+}
+
+
+ul li {
+	padding: 10px 0;
+}
+
+
+ul li:hover {
+	color: darkgreen;
+	cursor: pointer;
+}
+
+.online {
+	color: greenyellow;
+}
+
+.offline {
+	color: lightgray;
+}
