+++ client/src/components/ChatGroupsSideBar/index.js
@@ -1,0 +1,16 @@
+import React from 'react'
+import "./index.css"
+import ChatGroupsList from '../ChatGroupsList'
+import ChatGroupMembers from '../ChatGroupMembers'
+
+const ChatGroupsSideBar = () => {
+
+    return (
+        <aside className="chat-sidebar">
+            <ChatGroupsList />
+            <ChatGroupMembers />
+        </aside>
+    )
+}
+
+export default ChatGroupsSideBar
