+++ client/src/components/ChatProjectsSideBar/index.js
