+++ client/src/pages/ChatPage/index.js
@@ -1,12 +1,12 @@
 import React from 'react'
 import "./index.css"
-import ChatList from "../../components/ChatList"
-import ChatHeader from '../../components/ChatHeader'
-import ChatWindow from '../../components/ChatWindow'
-import ChatMessageInput from '../../components/ChatMessageInput'
-import ChatGroupMembers from '../../components/ChatGroupMembers'
 import SocketContextProvider from '../../context/SocketContext'
 import MessagesContextProvider from '../../context/MessagesContext'
+import ChatHeader from '../../components/ChatHeader'
+import ChatProjectsSideBar from "../../components/ChatProjectsSideBar"
+import ChatGroupsSideBar from '../../components/ChatGroupsSideBar'
+import ChatWindow from '../../components/ChatWindow'
+import ChatMessageInput from '../../components/ChatMessageInput'
 
 const ChatPage = () => {
 
@@ -16,8 +16,8 @@ const ChatPage = () => {
             <div className="chat-container">
                 <ChatHeader />
                 <main className="chat-main">
-                    <ChatList />
-                    <ChatGroupMembers />
+                    <ChatProjectsSideBar />
+                    <ChatGroupsSideBar />
                     <ChatWindow />
                 </main>
                 <ChatMessageInput />
